reorganize auth provider
